requirements.txt
transformers>=4.25.0
torch>=1.13.0
Pillow>=9.3.0
Flask>=2.2.0
flask-cors>=3.0.10
requests>=2.28.0
opencv-python-headless>=4.7.0